вернул модалку, убрал кэтчи из апи
