+++ src/components/burger-constructor/burger-constructor.tsx
@@ -19,6 +19,7 @@ import { useDrop } from "react-dnd/dist/hooks";
 import { IngredientObject } from "../../utils/interfaces";
 import { sendOrderRequest } from "../../services/constructorSlice";
 import ModalOverlay from "../modal-overlay/modal-overlay";
+import Modal from "../modal/modal";
 
 const BurgerConstructor = () => {
   // Достаем текущее состояние конструктора
@@ -149,9 +150,9 @@ const BurgerConstructor = () => {
       </main>
 
       {modalVisible && (
-        <ModalOverlay header={" "} onClosed={setModalVisible}>
+        <Modal header={" "} onClosed={setModalVisible}>
           <OrderDetails />
-        </ModalOverlay>
+        </Modal>
       )}
     </>
   );
