+++ src/components/modal-overlay/modal-overlay.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 import { ESC_BUTTON } from "../../utils/constants";
-import { ModalTypes } from "../../utils/interfaces";
-import Modal from "../modal/modal";
+import { ModalOverlayTypes } from "../../utils/interfaces";
 import modalOverlayStyles from "./modal-overlay.module.css";
 
-const ModalOverlay = ({ children, header, onClosed }: ModalTypes) => {
+const ModalOverlay = ({ children, onClosed }: ModalOverlayTypes) => {
   // Subscription on ESC button
   React.useEffect(() => {
     const close = (event: KeyboardEvent) => {
@@ -18,7 +17,7 @@ const ModalOverlay = ({ children, header, onClosed }: ModalTypes) => {
 
   return (
     <div className={modalOverlayStyles.overlay} onClick={onClosed}>
-      <Modal children={children} header={header} onClosed={onClosed} />
+      {children}
     </div>
   );
 };
